+++ bioblend/galaxy/users/__init__.py
@@ -1,7 +1,7 @@
 """
 Contains possible interaction dealing with Galaxy users.
 
-These methods must be executed by a registered Galaxy admin user.
+Most of these methods must be executed by a registered Galaxy admin user.
 """
 from bioblend.galaxy.client import Client
 
@@ -18,7 +18,7 @@ class UserClient(Client):
         get a list of deleted users.
 
         :rtype: list
-        :return: A list of dicts with user details.
+        :return: a list of dicts with user details.
                  For example::
 
                    [{u'email': u'a_user@example.com',
@@ -30,17 +30,16 @@ class UserClient(Client):
 
     def show_user(self, user_id, deleted=False):
         """
-        Display information about a user. If ``deleted`` is set to ``True``,
-        display information about a deleted user.
+        Display information about a user.
 
         :type user_id: str
-        :param user_id: User ID to inspect
+        :param user_id: encoded user ID
 
         :type deleted: bool
-        :param deleted: Whether to return results for a deleted user
+        :param deleted: whether to return results for a deleted user
 
         :rtype: dict
-        :return: dictionary containing information about the user
+        :return: a dictionary containing information about the user
         """
         return Client._get(self, id=user_id, deleted=deleted)
 
@@ -57,10 +56,10 @@ class UserClient(Client):
           to Galaxy via a browser will not be possible.
 
         :type user_email: str
-        :param user_email: Email of user to be created
+        :param user_email: email of the user to be created
 
         :rtype: dict
-        :return: dictionary containing information about the created user
+        :return: a dictionary containing information about the created user
         """
         payload = {}
         payload['remote_user_email'] = user_email
@@ -68,7 +67,7 @@ class UserClient(Client):
 
     def create_local_user(self, username, user_email, password):
         """
-        Create a new Galaxy user.
+        Create a new Galaxy local user.
 
         .. note::
           For this method to work, the Galaxy instance must have the
@@ -77,16 +76,16 @@ class UserClient(Client):
           ``config/galaxy.ini`` configuration file.
 
         :type username: str
-        :param username: Username of user to be created
+        :param username: username of the user to be created
 
         :type user_email: str
-        :param user_email: Email of user to be created
+        :param user_email: email of the user to be created
 
         :type password: str
-        :param password: password of user to be created
+        :param password: password of the user to be created
 
         :rtype: dict
-        :return: dictionary containing information about the created user
+        :return: a dictionary containing information about the created user
         """
         payload = {}
         payload['username'] = username
@@ -96,10 +95,11 @@ class UserClient(Client):
 
     def get_current_user(self):
         """
-        Returns the user id associated with this Galaxy connection
+        Display information about the user associated with this Galaxy
+        connection.
 
         :rtype: dict
-        :return: dictionary containing information about the current user
+        :return: a dictionary containing information about the current user
         """
         url = self.gi._make_url(self, None)
         url = '/'.join([url, 'current'])
@@ -107,13 +107,13 @@ class UserClient(Client):
 
     def create_user_apikey(self, user_id):
         """
-        Create a new api key for a user
+        Create a new API key for a given user.
 
         :type user_id: str
-        :param user_id: Encoded user ID
+        :param user_id: encoded user ID
 
         :rtype: str
-        :return: The api key for the user
+        :return: the API key for the user
         """
 
         url = self.gi._make_url(self, None)
@@ -122,3 +122,26 @@ class UserClient(Client):
         payload['user_id'] = user_id
 
         return Client._post(self, payload, url=url)
+
+    def delete_user(self, user_id, purge=False):
+        """
+        Delete a user.
+
+        .. note::
+          For this method to work, the Galaxy instance must have the
+          ``allow_user_deletion`` option set to ``True`` in the
+          ``config/galaxy.ini`` configuration file.
+
+        :type user_id: str
+        :param user_id: encoded user ID
+
+        :type purge: bool
+        :param purge: if ``True``, also purge (permanently delete) the history
+
+        :rtype: dict
+        :return: a dictionary containing information about the deleted user
+        """
+        params = {}
+        if purge is True:
+            params['purge'] = purge
+        return Client._delete(self, id=user_id, params=params)